incorporated log4j; updated emotivestock tests to run multiple times for randomized methods

+++ src/main/java/nmck/emotive_stocks/services/NYSE.java
@@ -3,6 +3,6 @@ package nmck.emotive_stocks.services;
 import java.time.LocalDate;
 
 public interface NYSE {
-    float getDailyGrowth(LocalDate localDate, String symbol);
+    double getDailyGrowth(LocalDate localDate, String symbol);
     boolean isMarketDay(LocalDate localDate);
 }
